+++ lib/Report.php
@@ -136,10 +136,10 @@ class Report implements \JsonSerializable {
     }
 
     catch (\Throwable $e) {
-      throw new \UnexpectedValueException('The coverage data has an invalid LCOV format');
+      throw new LcovException('The coverage data has an invalid LCOV format', $e);
     }
 
-    if (!count($records)) throw new \UnexpectedValueException('The coverage data is empty');
+    if (!count($records)) throw new LcovException('The coverage data is empty');
     return $report;
   }