+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Version [4.0.0](https://github.com/cedx/lcov.php/compare/v3.0.0...v4.0.0)
 - Breaking change: raised the required [PHP](https://secure.php.net) version.
 - Breaking change: using PHP 7.1 features, like nullable types and void functions.
+- Added a user guide based on [MkDocs](http://www.mkdocs.org).
+- Updated the package dependencies.
 
 ## Version [3.0.0](https://github.com/cedx/lcov.php/compare/v2.0.0...v3.0.0)
 - Breaking change: changed the signature of the data class constructors.