+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Version [4.1.0](https://github.com/cedx/lcov.php/compare/v4.0.0...v4.1.0)
+- Added the `LcovException` class.
+
 ## Version [4.0.0](https://github.com/cedx/lcov.php/compare/v3.0.0...v4.0.0)
 - Breaking change: raised the required [PHP](https://secure.php.net) version.
 - Breaking change: using PHP 7.1 features, like nullable types and void functions.